requirements.txt
# Core Data Science Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scikit-learn>=1.3.0,<1.6.0
scipy>=1.11.0,<1.14.0

# Deep Learning
tensorflow>=2.15.0,<2.18.0
keras>=3.0.0,<3.7.0

# Technical Indicators
ta>=0.11.0
pandas-ta>=0.3.14b

# Visualization
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0
plotly>=5.18.0,<6.0.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
python-dateutil>=2.8.2

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Streamlit GUI
streamlit>=1.28.0,<1.40.0
streamlit-option-menu>=0.3.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
